# Src/VicTrains.Application/TrainLines/Models/TrainLineListModel.cs
namespace VicTrains.Application.TrainLines.Models
{
    public class TrainLineListModel
    {
        public int Id { get; set; }
        public string Line { get; set; }
        public string Description { get; set; }
    }
}